actualizacion del sistema para lo de la integracion del ped publico de wm

+++ application/controllers/DocsPDF.php
@@ -13,6 +13,7 @@ class DocsPDF extends CI_Controller {
 
 	private $usuario;
 	private $default_pdf_params;
+	
 
 	function __construct()
 	{
@@ -116,12 +117,20 @@ class DocsPDF extends CI_Controller {
 				$doc_ped = $this->ArchivoModel->obtener_archivo($usuario_has_ec->id_archivo_ped);
 				$response['success'] = true;
 				$response['existe_ped'] = true;
+				$response['existe_ped_wm'] = false;
 				$response['msg'][] = array('');
 				$response['doc_portafolio_evidencia'] = base_url() . $doc_ped->ruta_directorio . $doc_ped->nombre;
+				//para el ped de wm
+				if(is_object($usuario_has_ec) && isset($usuario_has_ec->id_archivo_ped_wm) && !is_null($usuario_has_ec->id_archivo_ped_wm) && $usuario_has_ec->id_archivo_ped_wm != ''){
+					$doc_ped = $this->ArchivoModel->obtener_archivo($usuario_has_ec->id_archivo_ped_wm);
+					$response['existe_ped_wm'] = true;
+					$response['doc_portafolio_evidencia_wm'] = base_url() . $doc_ped->ruta_directorio . $doc_ped->nombre;
+				}
 			}else{
 				$data = $this->DocsPDFModel->get_data_portafolio_evidencia($id_usuario_alumno,$id_usuario_instructor,$id_estandar_competencia);
 				$response['success'] = true;
 				$response['existe_ped'] = false;
+				$response['existe_ped_wm'] = false;
 				$response['msg'][] = array('');
 				$response['data'] = $data;
 			}
@@ -156,9 +165,33 @@ class DocsPDF extends CI_Controller {
 				$mpdf->WriteHTML($paginaHTML);
 				$mpdf->Output(RUTA_PDF_TEMP.$nombre_documento, $output);
 			}
+
+			//agreamos el apartado para el ped que es de walmart
+			$nombre_documento_wm = $pre.'-01-portafolio_portada_wm.pdf';
+			if(!file_exists(RUTA_PDF_TEMP.$nombre_documento_wm)){
+				subdirectorios_files(RUTA_PDF_TEMP);
+				$data = $this->DocsPDFModel->get_data_portafolio_evidencia($id_usuario_alumno,$id_usuario_instructor,$id_estandar_competencia);
+				$this->default_pdf_params['margin_left'] = 0;
+				$this->default_pdf_params['margin_right'] = 0;
+				$this->default_pdf_params['margin_top'] = 0;
+				$this->default_pdf_params['margin_bottom'] = 0;
+				$mpdfWM = $this->pdf->load($this->default_pdf_params);
+				if(!es_produccion()){
+					$mpdfWM->SetWatermarkText('PED Demo - ECO SOFTyH');
+				}
+				$mpdfWM->showWatermarkText = true;
+				$paginaHTML = $this->load->view('pdf/portafolio_evidencia_wm/generar_portada', $data, true);
+				if($es_html != false && $es_html != 0){
+					echo $paginaHTML;exit;
+				}
+				$mpdfWM->WriteHTML($paginaHTML);
+				$mpdfWM->Output(RUTA_PDF_TEMP.$nombre_documento_wm, $output);
+			}
+
 			$response['success'] = true;
 			$response['msg'][] = 'Se genero el PDF del portafolio de evidencias (Portada hasta Ficha de registro)';
 			$response['data']['documento'] = RUTA_PDF_TEMP.$nombre_documento;
+			$response['data']['documento_wm'] = RUTA_PDF_TEMP.$nombre_documento_wm;
 		}catch (Exception $ex){
 			$response['success'] = false;
 			$response['msg'][] = 'Hubo un error en el sistema, intente nuevamente';
@@ -369,14 +402,15 @@ class DocsPDF extends CI_Controller {
 
 	public function generando_pdf_completo($id_usuario_alumno,$id_usuario_instructor,$id_estandar_competencia,$es_html = false,$output = 'F'){
 		try{
-			$mergePDF = new Merger();
 			$post = $this->input->post();
 			//$documentos_alumno_evidencia = $this->DocsPDFModel->obtener_archivos_ec_alumno($id_usuario_alumno,$id_estandar_competencia);
 			//nueva consulta de información de la evidencia de los candidatos
 			$documentos_alumno_evidencia = $this->DocsPDFModel->obtener_archivos_ec_alumno_entregables($id_usuario_alumno,$id_estandar_competencia);
 			$data = $this->DocsPDFModel->get_data_portafolio_evidencia($id_usuario_alumno,$id_usuario_instructor,$id_estandar_competencia);
 			$archivos_conjuntar = array();
+			$archivos_conjuntar_ped_wm = array();
 			$archivos_conjuntar[] = $post['docs_generados'][0];//portada a antes de la ficha de registro
+			$archivos_conjuntar_ped_wm[] = $post['docs_generados_wm'][0]; //portada
 			if(!es_produccion()){
 				//ficha registro cargado al sistema por el admin/evaluador y se le pone la marca de agua del sistema
 				$archivos_conjuntar[] = $this->poner_marca_agua_doc($data['usuario_has_expediente_ped'][0]->ruta_directorio,$data['usuario_has_expediente_ped'][0]->nombre);
@@ -394,28 +428,55 @@ class DocsPDF extends CI_Controller {
 			//entregables candidato
 			$archivos_entregables_modificados = $this->generar_entregables($documentos_alumno_evidencia);
 			$archivos_conjuntar = array_merge($archivos_conjuntar,$archivos_entregables_modificados);
+			$archivos_conjuntar_ped_wm = array_merge($archivos_conjuntar_ped_wm,$archivos_entregables_modificados);
+
 			$archivos_conjuntar[] = $post['docs_generados'][3];//cierre hasta antes del certificado
+			
 			//certificado del conocer
-			$archivos_conjuntar[] = $this->certificado_formato_ped(
+			$certificado_conocer = $this->certificado_formato_ped(
 				$data['foto_firma']->ruta_directorio.$data['foto_firma']->nombre,
 				$data['usuario_alumno']->nombre.' '.$data['usuario_alumno']->apellido_p.' '.$data['usuario_alumno']->apellido_m,
 				$data['usuario_has_expediente_ped'][2] //certificado conocer original
 			);
+
+			$archivos_conjuntar[] = $certificado_conocer;
+			$archivos_conjuntar_ped_wm[] = $certificado_conocer;
+			$constancia_dc3 = $this->constancia_dc3($data['usuario_has_ec']->id_usuario_has_estandar_competencia,true);
+			$archivos_conjuntar_ped_wm[] = $constancia_dc3->ruta_directorio.$constancia_dc3->nombre;
 			$archivos_conjuntar[] = 'assets/docs/final_ped.pdf';
-			//conjuntamos los archivos generados previamente
+			$archivos_conjuntar_ped_wm[] = 'assets/docs/final_ped.pdf';
+			//conjuntamos los archivos generados previamente para el ped
+			$mergePDF = new Merger();
 			foreach ($archivos_conjuntar as $dg){
+				log_message('error',$dg);
 				$mergePDF->addFile(FCPATH.$dg);
 			}
 
 			$archivo_combinado = $mergePDF->merge();
+			//var_dump($archivo_combinado);exit;
 			$pre = date('Ymd').'-'.$id_usuario_alumno.'-'.$id_estandar_competencia;
 			$nombre_documento = $pre.'-final-portafolio-evidencias.pdf';
 			$directorio = $this->subdirectorio_peds_generados();
 			subdirectorios_files($directorio);
 			$isArchivo_generado = file_put_contents($directorio.$nombre_documento,$archivo_combinado);
+
+			//conjuntamos los archivos generados para el ped de walmart
+			$mergePDFPEDWM = new Merger();
+			foreach($archivos_conjuntar_ped_wm as $dg){
+				log_message('error',$dg);
+				$mergePDFPEDWM->addFile(FCPATH.$dg);
+			}
+			
+			$archivo_combinado_ped_wm = $mergePDFPEDWM->merge();
+			$pre = date('Ymd').'-'.$id_usuario_alumno.'-'.$id_estandar_competencia;
+			$nombre_documento_wm = $pre.'-final-portafolio-evidencias-wm.pdf';
+			$directorio = $this->subdirectorio_peds_generados();
+			subdirectorios_files($directorio);
+			$isArchivo_generado_wm = file_put_contents($directorio.$nombre_documento_wm,$archivo_combinado_ped_wm);
+
 			if($isArchivo_generado){
 				foreach ($archivos_conjuntar as $dg){
-					if(strpos($dg,RUTA_PDF_TEMP) !== false){
+					if(strpos($dg,RUTA_PDF_TEMP) !== false && file_exists(FCPATH.$dg)){
 						unlink(FCPATH.$dg);
 					}
 				}
@@ -427,17 +488,42 @@ class DocsPDF extends CI_Controller {
 				$this->DocsPDFModel->actualizar_ped_generado($id_usuario_alumno,$id_estandar_competencia,$id_archivo);
 				$response['success'] = true;
 				$response['msg'][] = 'Se genero el Portafolio de evidencias del alumno correctamente';
-				$response['data'] = $datos_doc;
+				$response['data']['ped'] = $datos_doc;
 			}else{
 				$response['success'] = false;
 				$response['msg'][] = 'Hubo un error en el sistema, intente nuevamente';
 			}
+
+			if($isArchivo_generado_wm){
+				foreach ($archivos_conjuntar_ped_wm as $dg){
+					if(strpos($dg,RUTA_PDF_TEMP) !== false && file_exists(FCPATH.$dg)){
+						unlink(FCPATH.$dg);
+					}
+				}
+				//almacenamos el archivo en la tabla de archivo y luego en la tabla de referencia de que se genero el PED
+				$datos_doc['nombre'] = $nombre_documento_wm;
+				$datos_doc['ruta_directorio'] = $directorio;
+				$datos_doc['fecha'] = date('Y-m-d H:i:s');
+				$id_archivo = $this->ArchivoModel->guardar_archivo_model($datos_doc);
+				$this->DocsPDFModel->actualizar_ped_wm_generado($id_usuario_alumno,$id_estandar_competencia,$id_archivo);
+				$response['success'] = true;
+				$response['msg'][] = 'Se genero el Portafolio de evidencias del alumno correctamente';
+				$response['data']['ped_wm'] = $datos_doc;
+			}else{
+				$response['success'] = false;
+				$response['msg'][] = 'Hubo un error en el sistema, intente nuevamente';
+			}
+
 		}catch (Exception $ex){
 			$response['success'] = false;
 			$response['msg'][] = 'Hubo un error en el sistema, intente nuevamente';
 			$response['msg'][] = $ex->getMessage();
 		}
 		echo json_encode($response);exit;
+	}
+
+	public function generando_pdf_wm_completo(){
+
 	}
 
 
@@ -577,6 +663,176 @@ class DocsPDF extends CI_Controller {
 			log_message('error',$ex->getMessage());
 		}
 		echo 'Ocurrio un error en el proceso de generar la firma, intente más tarde';
+	}
+
+	public function constancia_dc3($id_usuario_has_estandar_competencia,$para_ped_wm = false){
+		try{
+			$this->load->model('EstandarCompetenciaModel');
+			$this->load->model('EstandarCompetenciaGrupoModel');
+			$this->load->model('PerfilModel');
+			//apartado de los datos del candidato para la dc3
+			
+			//traemos los datos para poder generar la credencial y codigoqr
+			$usuario_has_estandar_competencia = $this->UsuarioHasECModel->obtener_row($id_usuario_has_estandar_competencia);
+			//validamos que ya se pueda emitir la credencial con una calificacion aprobatoria
+			if((int)$usuario_has_estandar_competencia->id_cat_calibracion_desempeno < JUICIO_CALIFICADO){
+				echo 'Candidato aun no calificado';exit;
+			}
+
+			//validamos si existe la dc3 de este usuario y el estandar de competencia
+			if(isset($usuario_has_estandar_competencia->id_archivo_dc3) && !is_null($usuario_has_estandar_competencia->id_archivo_dc3)){
+				$constanciaDC3 = $this->ArchivoModel->obtener_archivo($usuario_has_estandar_competencia->id_archivo_dc3);
+				if($para_ped_wm){
+					return $constanciaDC3;
+				}else{
+					redirect(base_url().$constanciaDC3->ruta_directorio.$constanciaDC3->nombre);
+				}
+			}else{
+				$datos_usuario = $this->UsuarioModel->obtener_data_usuario_id($usuario_has_estandar_competencia->id_usuario);
+				$datos_instructor = $this->UsuarioModel->obtener_data_usuario_id($usuario_has_estandar_competencia->id_usuario_evaluador);
+				$firma_instructor = $this->PerfilModel->obtener_datos_expediente($usuario_has_estandar_competencia->id_usuario_evaluador,EXPEDIENTE_FIRMA_DIGITAL);
+				$datos_empresa = $this->PerfilModel->obtener_datos_empresa($usuario_has_estandar_competencia->id_usuario,true);
+				$cat_ocupacion_especifica = $this->CatalogoModel->get_catalogo('cat_ocupacion_especifica',$datos_usuario->id_cat_ocupacion_especifica);
+				$estandar_competencia = $this->EstandarCompetenciaModel->obtener_row($usuario_has_estandar_competencia->id_estandar_competencia);
+				$estandar_competencia_grupo = $this->EstandarCompetenciaGrupoModel->obtener_row($usuario_has_estandar_competencia->id_estandar_competencia_grupo);
+				$cat_area_tematica = $this->CatalogoModel->get_catalogo('cat_area_tematica',$estandar_competencia_grupo->id_cat_area_tematica);
+				
+				$pdf = new Fpdi();
+				$mpdf = $this->pdf->load($this->default_pdf_params);
+				//leemos la plantilla para generar el GAFETE
+				$pdf->setSourceFile(RUTA_PLANTILLA_CONSTANCIA_DC3);
+				
+				$paginaActual = $pdf->importPage(1);
+				$paPlantilla = $pdf->getTemplatesize($paginaActual);
+				$pdf->AddPage($paPlantilla['orientation'],$paPlantilla);
+				$pdf->useImportedPage($paginaActual);
+				//agregamos la marca de agua para pruebas
+				if(!es_produccion()) {
+					//agregamos el texto
+					$pdf->SetFont('Arial','',40);
+					$pdf->SetTextColor(150, 150, 150);
+					$pdf->SetXY(20, $paPlantilla[1] / 2);
+					$pdf->Write(0, utf8_decode('PED Demo - ECO SOFTyH'));
+				}
+
+				//agregamos los logotipos de la constancia
+				//imagen de la empresa del candidato
+				$pdf->Image(FCPATH.$datos_empresa->ruta_directorio_logo.$datos_empresa->nombre_archivo_logo,10,5,60,20);
+				//logotipo civika
+				$pdf->Image(FCPATH.RUTA_LOGO_CIVIKA_PDF,145,5,60,20);
+
+				//nombre del trabajador
+				$pdf->SetFont('Arial','B',10);
+				$pdf->SetTextColor(0,0,0);
+				$pdf->SetXY(10,64);
+				$pdf->Write(0,utf8_decode($datos_usuario->apellido_p.' '.$datos_usuario->apellido_m.' '.$datos_usuario->nombre));
+				//curp
+				$curp = str_split($datos_usuario->curp);
+				$posXCurp = [10,15.5,21.5,27, 33,38.5,44,49, 55,60,65.5,71.4, 76.6,82,88,93.5, 99,105];
+				foreach($curp as $index => $letra_curp){
+					$pdf->SetXY($posXCurp[$index],76);
+					$pdf->Write(0,utf8_decode($letra_curp));
+				}
+				//ocupacion especifica
+				$pdf->SetXY(110.5,76);
+				$pdf->Write(0,utf8_decode($cat_ocupacion_especifica->clave_area_subarea.' - '. $cat_ocupacion_especifica->denominacion));
+				//puesto
+				$pdf->SetXY(10,85);
+				$pdf->Write(0,utf8_decode($datos_empresa->cargo));
+				
+				//nombre de la empresa
+				$pdf->SetXY(10,106);
+				$pdf->Write(0,utf8_decode($datos_empresa->nombre));
+				//rfc empresa
+				$rfc_empresa = str_split($datos_empresa->rfc);
+				if(sizeof($rfc_empresa) == 12){
+					$posXRfc = [15,20.5,26.2, 37,42.5,48,53,59,64, 75,80,86 ];
+				}else{
+					$posXRfc = [10,15,20.5,26.2, 37,42.5,48,53,59,64, 75,80,86 ];
+				}
+				foreach($rfc_empresa as $index => $letra_rfc){
+					$pdf->SetXY($posXRfc[$index],117.5);
+					$pdf->Write(0,utf8_decode($letra_rfc));
+				}
+
+				//estandar de competencia
+				$pdf->SetXY(10,137);
+				$pdf->Write(0,utf8_decode($estandar_competencia->codigo.' - '.$estandar_competencia->titulo));
+				//duracion en horas
+				$pdf->SetXY(10,148);
+				$pdf->Write(0,utf8_decode($estandar_competencia_grupo->duracion));
+				//periodo de ejecucion
+				$periodo_ejecucion = str_split(str_replace('-','',$estandar_competencia_grupo->periodo_inicio.$estandar_competencia_grupo->periodo_fin));
+				$posXPeriodo = [91.5,98,105,112, 119,126, 133,140, 153,160,167,174, 181,188, 195,201.5 ];
+				foreach($periodo_ejecucion as $index => $letra_periodo){
+					$pdf->SetXY($posXPeriodo[$index],148);
+					$pdf->Write(0,$letra_periodo);
+				}
+
+				//area tematica
+				$pdf->SetXY(10,158.5);
+				$pdf->Write(0,utf8_decode($cat_area_tematica->clave.' - '.$cat_area_tematica->area_tematica));
+				//agente capacitador
+				$pdf->SetXY(10,169);
+				$pdf->Write(0,utf8_decode($estandar_competencia_grupo->agente_capacitador));
+
+				//nombre y firma instructor
+				$pdf->SetXY(15,207);
+				$pdf->Write(0,utf8_decode($datos_instructor->apellido_p.' '.$datos_instructor->apellido_m.' '.$datos_instructor->nombre));
+				$pdf->Image(FCPATH.$firma_instructor->ruta_directorio.$firma_instructor->nombre,25,185,35,18);
+
+				//sello
+				$pdf->Image(FCPATH.RUTA_SELLO_CIVIKA_PDF,62,185);
+
+				//representante legal
+				$pdf->SetXY(82,207);
+				$pdf->Write(0,utf8_decode($datos_empresa->representante_legal));
+				//representante de trabajadores
+				$pdf->SetXY(147,207);
+				$pdf->Write(0,utf8_decode($datos_empresa->representante_trabajadores));
+
+				//el QR para la constancia dc3
+				if(isset($usuario_has_estandar_competencia->id_archivo_dc3) && !is_null($usuario_has_estandar_competencia->id_archivo_dc3)){
+					$codigoQRDC3 = $this->ArchivoModel->obtener_archivo($usuario_has_estandar_competencia->id_archivo_dc3);
+				}else{
+					//en caso de que no exista el codigo QR para el candidato se genera uno nuevo
+					//el codigo qr es generico o general para el usuario ya que al escanear se manda a un perfil publico
+					$qrGenerado = $this->generar_qr_dc3_usuario_candidato($id_usuario_has_estandar_competencia);
+					if($qrGenerado['success']){
+						$codigoQRDC3 = $qrGenerado['data']['archivo_qr'];
+					}
+				}
+
+				$pdf->SetFont('Arial','B',6);
+				$pdf->SetXY(184,249);
+				$folio = 'Folio: '.str_replace('.png','',$codigoQRDC3->nombre);
+				$pdf->Write(0,utf8_decode($folio));
+				$pdf->Image(FCPATH.$codigoQRDC3->ruta_directorio.$codigoQRDC3->nombre,184,222,24,24);
+
+				$directorio = RUTA_PDF_PED.date('Y').'/'.date('m').'/'.date('d').'/';
+				subdirectorios_files($directorio);
+				$datos_doc['nombre'] = 'CONSTANCIA-DC3-'.str_replace('.png','',$codigoQRDC3->nombre);
+				$datos_doc['ruta_directorio'] = $directorio;
+				$datos_doc['fecha'] = date('Y-m-d H:i:s');
+				
+				$pdf->Output('F', FCPATH.$datos_doc['ruta_directorio'].$datos_doc['nombre']);
+				$pdf->cleanUp();
+
+				$id_archivo = $this->ArchivoModel->guardar_archivo_model($datos_doc);
+				$this->UsuarioHasECModel->guardar_row(array('id_archivo_dc3' => $id_archivo),$id_usuario_has_estandar_competencia);
+				if($para_ped_wm){
+					return $this->ArchivoModel->obtener_archivo($id_archivo);
+				}else{
+					redirect(base_url().$datos_doc['ruta_directorio'].$datos_doc['nombre']);
+					echo 'Se genero la constancia DC3 correctamente';	
+				}
+			}
+		}catch (Exception $ex){
+			log_message('error','***** DocsPDFModel -> certificado_dc3');
+			log_message('error',$ex->getMessage());
+			log_message('error',$ex->getFile().' - '.$ex->getLine());
+			echo 'Ocurrio un error en el proceso de generar la constancia DC3- favor de intentar mas tarde';
+		}
 	}
 
 	protected function set_variables_defaults_pdf()
@@ -840,6 +1096,45 @@ class DocsPDF extends CI_Controller {
 			$datos_doc['fecha'] = date('Y-m-d H:i:s');
 			$id_archivo = $this->ArchivoModel->guardar_archivo_model($datos_doc);
 			$this->UsuarioModel->guardar_row(array('id_archivo_qr' => $id_archivo),$idUsuario);
+			$archivoQr = new stdClass();
+			$archivoQr->id_archivo = $id_archivo;
+			$archivoQr->nombre = $qr_image;
+			$archivoQr->ruta_directorio = $directorio;
+			$archivoQr->fecha = $datos_doc['fecha'];
+			$response['data']['archivo_qr'] = $archivoQr;
+		}catch (Exception $ex){
+			$response['success'] = false;
+			$response['msg'][] = 'Hubo un error en el sistema, intente nuevamente';
+			$response['msg'][] = $ex->getMessage();
+			log_message('error','***** DocsPDFModel -> generar_qr_usuario_candidato');
+			log_message('error',$ex->getMessage());
+		}
+		return $response;
+	}
+
+	protected function generar_qr_dc3_usuario_candidato($id_usuario_has_estandar_competencia){
+		try{
+			$response['success'] = true;
+			$response['msg'][] = 'Se genero el QR correctamente';
+			$stringToQR = base_url().'constancia_dc3/'.$id_usuario_has_estandar_competencia;
+			//var_dump('aqui toy');exit;
+			$this->load->library('ciqrcode');
+			$directorio = RUTA_QR_FILES_PERFIL.date('Y').'/'.date('m').'/'.date('d').'/';
+			subdirectorios_files($directorio);
+			$qr_image = date('Ymd').'-'.$id_usuario_has_estandar_competencia.'.png';
+			$params['data'] = $stringToQR;
+			$params['level'] = 'l';
+			$params['size'] = 300;
+			$params['savename'] = FCPATH.$directorio.$qr_image;
+			if(!file_exists($params['savename'])){
+				$this->ciqrcode->generate($params);
+			}
+			//guardamo el archivo generado y lo almacenamos en la tabla de archivo
+			$datos_doc['nombre'] = $qr_image;
+			$datos_doc['ruta_directorio'] = $directorio;
+			$datos_doc['fecha'] = date('Y-m-d H:i:s');
+			$id_archivo = $this->ArchivoModel->guardar_archivo_model($datos_doc);
+			$this->UsuarioHasECModel->guardar_row(array('id_archivo_dc3_qr' => $id_archivo),$id_usuario_has_estandar_competencia);
 			$archivoQr = new stdClass();
 			$archivoQr->id_archivo = $id_archivo;
 			$archivoQr->nombre = $qr_image;
